Uses a base class for subtitlesettingsbutton
