+++ src/ts/components/subtitlesettings/subtitlesettingsclosebutton.ts
@@ -1,23 +1,17 @@
-import {Button, ButtonConfig} from '../button';
+import {SubtitleSettingsButton} from './subtitlesettingsbutton';
 import {SettingsPanel} from '../settingspanel';
 import {SubtitleSettingsPanel} from '../subtitlesettingspanel';
-import {SubtitleSettingsButtonConfig} from '../subtitlesettingtoggle';
+import {SubtitleSettingsButtonConfig} from './subtitlesettingsbutton';
 import {UIInstanceManager} from '../../uimanager';
 import {DOM} from '../../dom';
 
 /**
  * A button that toggles the option menu for subtitles
  */
-export class SubtitleSettingsCloseButton extends Button<ButtonConfig> {
-
-  private subtitleSettingsPanel: SubtitleSettingsPanel;
-  private settingsPanel: SettingsPanel;
+export class SubtitleSettingsCloseButton extends SubtitleSettingsButton {
 
   constructor(config: SubtitleSettingsButtonConfig) {
     super(config);
-
-    this.subtitleSettingsPanel = config.subtitleSettingsPanel;
-    this.settingsPanel = config.settingsPanel;
 
     this.config = this.mergeConfig(config, {
       cssClass: 'ui-subtitleoptioncloser',
